app/pyproject.toml
[tool.poetry]
name = "churn"
version = "0.1.0"
description = ""
authors = ["Deviyanti Aryani <deviyantiam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
db-dtypes = "1.2.0"
google-cloud-bigquery = "3.20.1"
numpy = "1.24.4"
pandas = "2.0.3"
scikit-learn = "1.3.2"
python-dotenv = "^1.0.1"
tornado = "6.0.4"
joblib = "^1.3.2"
pyarrow = "^15.0.2"
tqdm = "^4.66.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
